This was asked in a Hacker rank test
This was asked in an assessment exam called the PRA of an undisclosed company name. I wasn't able to complete the code within time. Absolutely gutted afterward that I couldn't clear the assessment. But, still happy that I was able to pull something out on my own. I needed just another 20 seconds to submit this code which ran out. Glad I took that little effort to complete the code afterward and share it with you guys.